+++ src/app/dashboard/matkul/page.tsx
@@ -18,6 +18,7 @@ export default async function MatkulListPage() {
     dosens: { select: { id: true, name: true } },
     koordinators: { select: { id: true, name: true } },
     semester: { include: { tahunAkademik: { select: { tahun: true } } } },
+    classes: { select: { id: true, name: true } },
   };
 
   const [dosenMatkuls, koordinatorMatkuls, semesters] = await Promise.all([